qml: add a special grid view where elements can be expanded
ExpandGridView mimics the behavior of a QML Grid view except it allows to
display an Item below the currently selected item.

+-----------------+
| [ ] [ ] [X] [ ] | <- [X] selected item of the grid
| --------------- |
|                 | <- item details here
| --------------- |
| [ ] [ ] [ ] [ ] | <- next items of the grid
+-----------------+

The implementation is made by using two GridView and handling position and
content position according to the global position of the view.

The main drawback is that the model/delegate to be rendered in the grid has to
be passed twice has ModelDelegate can't be shared between two gridview

Signed-off-by: Thomas Guillem <thomas@gllm.fr>